# server/swagger.yml
openapi: "3.0.0"
info:
  title: "E-Commerce Platform API"
  version: "1.0.0"
  description: |
    # E-Commerce REST API Documentation
    
    Modern full-stack e-commerce API with Domain-Driven Design architecture.
    
    ## API Overview
    - **Base URL**: `/api/v1`
    - **Authentication**: Bearer token (JWT)
    - **Rate Limiting**: Applied to auth and payment endpoints
    - **Response Format**: JSON
    
    ## Domain Routes
    - **Authentication** (`/auth`) - User registration, login, token management
    - **Identity** (`/identity`) - User profiles, addresses, payment methods
    - **Catalog** (`/catalog`) - Products, categories, wishlist
    - **Ordering** (`/ordering`) - Cart, orders, checkout, payments, shipping
    - **Vendors** (`/vendors`) - Vendor management and applications
    
  license:
    name: "MIT"
    url: "https://opensource.org/licenses/MIT"
servers:
  - url: "http://localhost:5000"
    description: "Development"
  - url: "https://staging.ecommerce.example"
    description: "Staging"
  - url: "https://api.ecommerce.example"
    description: "Production"
security:
  - bearerAuth: []
paths:
...
  /api/v1/auth/change-password:
    post:
      operationId: changePassword
      summary: "Change user password"
      description: "Change password with current password verification"
      tags: ["Authentication"]
      responses:
        "200":
          description: "Password changed successfully"
        "401":
          description: "Current password is incorrect"
...
  /api/v1/identity/profile:
    get:
      operationId: getProfile
      summary: "Get user profile"
      description: "Retrieve authenticated user's profile information"
      tags: ["Identity"]
      responses:
        "200":
          description: "Profile retrieved successfully"
        "401":
          description: "Unauthorized"
    put:
      operationId: updateProfile
      summary: "Update user profile"
      description: "Update user's profile information"
      tags: ["Identity"]
      responses:
        "200":
          description: "Profile updated successfully"
        "401":
          description: "Unauthorized"
    delete:
      operationId: deleteAccount
      summary: "Delete user account"
      description: "Permanently delete user account"
      tags: ["Identity"]
      responses:
        "200":
          description: "Account deleted successfully"
        "401":
          description: "Unauthorized"
  /api/v1/identity/profile/addresses:
    get:
      operationId: getAddresses
      summary: "Get all user addresses"
      description: "Retrieve all addresses for authenticated user"
      tags: ["Identity"]
      responses:
        "200":
          description: "Addresses retrieved successfully"
        "401":
          description: "Unauthorized"
    post:
      operationId: addAddress
      summary: "Add new address"
      description: "Add a new shipping or billing address"
      tags: ["Identity"]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddressInput"
      responses:
        "201":
          description: "Address added successfully"
        "401":
          description: "Unauthorized"
  /api/v1/identity/profile/addresses/{addressId}:
    patch:
      operationId: updateAddress
      summary: "Update address"
      description: "Update an existing address"
      tags: ["Identity"]
      parameters:
        - name: addressId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddressInput"
      responses:
        "200":
          description: "Address updated successfully"
        "401":
          description: "Unauthorized"
    delete:
      operationId: deleteAddress
      summary: "Delete address"
      description: "Delete an address"
      tags: ["Identity"]
      parameters:
        - name: addressId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: "Address deleted successfully"
        "401":
          description: "Unauthorized"
  /api/v1/identity/profile/cards:
    get:
      operationId: getSavedCards
      summary: "Get all saved cards"
      description: "Retrieve all saved payment cards"
      tags: ["Identity"]
      responses:
        "200":
          description: "Saved cards retrieved successfully"
        "401":
          description: "Unauthorized"
    post:
      operationId: addSavedCard
      summary: "Add saved card"
      description: "Save a new payment card for future use"
      tags: ["Identity"]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        "201":
          description: "Card saved successfully"
        "401":
          description: "Unauthorized"
  /api/v1/identity/profile/cards/{cardId}:
    delete:
      operationId: deleteSavedCard
      summary: "Delete saved card"
      description: "Delete a saved payment card"
      tags: ["Identity"]
      parameters:
        - name: cardId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: "Card deleted successfully"
        "401":
          description: "Unauthorized"
  /api/v1/identity/profile/cards/{cardId}/primary:
    patch:
      operationId: setPrimaryCard
      summary: "Set primary card"
      description: "Set a card as the primary payment method"
      tags: ["Identity"]
      parameters:
        - name: cardId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: "Primary card updated successfully"
        "401":
          description: "Unauthorized"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  schemas:
...
    AddressInput:
      type: object
      required:
        - type
        - street
        - city
        - country
      properties:
        type:
          type: string
          enum: ["shipping", "billing"]
        street:
          type: string
        city:
          type: string
        state:
          type: string
        country:
          type: string
        postalCode:
          type: string